requirements.txt
requests~=2.32.5
mcp~=1.12.1
python-dotenv~=1.1.0